refactor: Rename ProviderManager to TenantProviderManager to keep consistency
