+++ tests/Unit/SproutServiceProviderTest.php
@@ -13,7 +13,7 @@ use Sprout\Events\CurrentTenantChanged;
 use Sprout\Http\Middleware\TenantRoutes;
 use Sprout\Listeners\IdentifyTenantOnRouting;
 use Sprout\Managers\IdentityResolverManager;
-use Sprout\Managers\ProviderManager;
+use Sprout\Managers\TenantProviderManager;
 use Sprout\Managers\TenancyManager;
 use Sprout\Sprout;
 use Sprout\SproutServiceProvider;
@@ -54,17 +54,17 @@ class SproutServiceProviderTest extends UnitTestCase
     #[Test]
     public function providerManagerIsRegistered(): void
     {
-        $this->assertTrue(app()->has(ProviderManager::class));
+        $this->assertTrue(app()->has(TenantProviderManager::class));
         $this->assertTrue(app()->has('sprout.providers'));
-        $this->assertTrue(app()->isShared(ProviderManager::class));
+        $this->assertTrue(app()->isShared(TenantProviderManager::class));
         $this->assertFalse(app()->isShared('sprout.providers'));
 
-        $this->assertSame(app()->make(ProviderManager::class), app()->make(ProviderManager::class));
+        $this->assertSame(app()->make(TenantProviderManager::class), app()->make(TenantProviderManager::class));
         $this->assertSame(app()->make('sprout.providers'), app()->make('sprout.providers'));
-        $this->assertSame(app()->make(ProviderManager::class), app()->make('sprout.providers'));
-        $this->assertSame(app()->make('sprout.providers'), app()->make(ProviderManager::class));
+        $this->assertSame(app()->make(TenantProviderManager::class), app()->make('sprout.providers'));
+        $this->assertSame(app()->make('sprout.providers'), app()->make(TenantProviderManager::class));
         $this->assertSame(app()->make(Sprout::class)->providers(), app()->make('sprout.providers'));
-        $this->assertSame(app()->make(Sprout::class)->providers(), app()->make(ProviderManager::class));
+        $this->assertSame(app()->make(Sprout::class)->providers(), app()->make(TenantProviderManager::class));
         $this->assertSame(sprout()->providers(), sprout()->providers());
         $this->assertSame(app()->make(Sprout::class)->providers(), sprout()->providers());
     }
